fix(auth): fix decodeIdToken type and fast refresh warnings
- Type decodeIdToken as User | null; it returns null on decode failure
  but was annotated User, causing a TS2322 error.
- Split AuthContext into context.ts (context + useAuth hook) and
  AuthProvider.tsx (component) with a barrel index, so the module
  exports only components and the react-refresh/only-export-components
  warnings clear.
- Extract the shared base64url JWT decode into decodeJwtPayload,
  removing the duplicate block in decodeIdToken and isTokenExpired.
- Bump version to 1.10.2.
