+++ src/contexts/AuthContext/index.ts
@@ -1,0 +1,2 @@
+export { AuthContext, useAuth, type AuthContextValue } from './context'
+export { AuthProvider } from './AuthProvider'
